fix regulation checks in CI
